pyproject.toml
[project]
name = "finsightai-stramlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "st-social-media-links>=0.1.5",
    "streamlit>=1.46.1",
]
